+++ lib/plugins/dtf/stats_output.rb
@@ -5,6 +5,10 @@ class DTF::StatsOutput
   BLUE = `tput setaf 4`
   RESET = `tput setaf 9`
 
+  def self.argument_matches? argument
+    [:load] if argument == "--text"
+  end
+
   def initialize
     @counts={}
     @counts[:commands] = 0