+++ test/unit/error_summary_output.rb
@@ -1,9 +1,18 @@
 require 'minitest/autorun'
-require 'plugins/dtf/stats_output'
+require 'plugins/dtf/error_summary_output'
 
-class TestStatsOutput < MiniTest::Unit::TestCase
+class TestErrorSummaryOutput < MiniTest::Unit::TestCase
   def setup
-    @test = DTF::StatsOutput.new
+    @test = DTF::ErrorSummaryOutput.new self
+    @outputs = []
+  end
+
+  def printf args
+    @outputs << args
+  end
+
+  def puts args
+    @outputs << args
   end
 
   # Build input data, will create command for every integer passed,
@@ -28,6 +37,7 @@ class TestStatsOutput < MiniTest::Unit::TestCase
 
   def test_processed_success
     @test.start_test(build_input(1), {})
+    @test.start_command( {} )
     @test.end_command( '', true, {} )
     @test.test_processed( {}, true, '' )
 
@@ -38,16 +48,18 @@ class TestStatsOutput < MiniTest::Unit::TestCase
 
   def test_processed_failure
     @test.start_test(build_input(1), {})
-    @test.end_command( '', true, {} )
+    @test.start_command( {} )
+    @test.end_command( {}, true, {} )
     @test.test_processed( {}, false, '' )
 
-    assert_match( /##### Processed commands 1 of 1/, @test.status )
-    assert_match( /failure tests 1 of 1/, @test.status )
-    refute_match( /(successskipped)/, @test.status )
+   assert_match( /##### Processed commands 1 of 1/, @test.status )
+   assert_match( /failure tests 1 of 1/, @test.status )
+   refute_match( /(successskipped)/, @test.status )
   end
 
   def test_processed_skipped
     @test.start_test(build_input(1), {})
+    @test.start_command( {} )
     @test.end_command( '', true, {} )
 
     assert_match( /##### Processed commands 1 of 1/, @test.status )
@@ -55,9 +67,13 @@ class TestStatsOutput < MiniTest::Unit::TestCase
     refute_match( /(success|failure)/, @test.status )
   end
 
+  #TODO: test @outputs
   def test_processed_all
     @test.start_test(build_input(1,2), {})
-    2.times{ @test.end_command('',true,{}) }
+    2.times{
+      @test.start_command( {} )
+      @test.end_command('',true,{})
+    }
     @test.test_processed( {}, true, '' )
     @test.test_processed( {}, false, '' )